Break TwitterPreview component into smaller TS components
