+++ packages/social-previews/src/twitter-preview/sidebar.tsx
@@ -1,0 +1,13 @@
+import { __ } from '@wordpress/i18n';
+import { SidebarProps } from './types';
+
+export const Sidebar: React.FC< SidebarProps > = ( { profileImage, isLast } ) => {
+	return (
+		<div className="twitter-preview__sidebar">
+			<div className="twitter-preview__profile-image">
+				<img alt={ __( 'Twitter profile image' ) } src={ profileImage } />
+			</div>
+			{ ! isLast && <div className="twitter-preview__connector" /> }
+		</div>
+	);
+};
